Revert "updated detectType for rhino functions"
