+++ lib/rhyton/utils.py
@@ -6,9 +6,6 @@ Module for general utily functions.
 # rhyton imports
 from rhyton.main import Rhyton
 
-# rhino imports
-import Rhino
-import rhinoscriptsyntax as rs
 
 class Format:
     """
@@ -158,7 +155,7 @@ def removePrefix(string, prefix):
 
 def detectType(value):
     """
-    Tries to convert a given string to a number, boolean or string or function.
+    Tries to convert a given string to a number, boolean or string.
 
     Args:
         value (str): The string to convert.
@@ -182,35 +179,4 @@ def detectType(value):
         return False
 
     return value
-
-def _detect_function_type(uTxt, objID):
-    """
-    Detects if the value is calculated with a Rhino function
-
-    Args:
-        value (str): the string check
-    Returns:
-        value: converted or original value
-    """
-    if not uTxt:
-        return uTxt
-    if uTxt[:2] == "%<" and uTxt[-2:] == ">%":
-        obj = rs.coercerhinoobject(objID)
-        uTxt = Rhino.RhinoApp.ParseTextField(uTxt, obj, None)
-    return uTxt
-
-def getValue(guid, key):
-    """
-    Wrapper function to get user text from an object
-
-    Args:
-        guid (str): A rhino objects id.
-        key (str): The key to get the value from.
-
-    Returns:
-        mixed: None if key does not exist,
-        " " if key has no value,
-        else: str of value
-    """
-    keyValue = _detect_function_type(rs.GetUserText(guid, key), guid)
-    return detectType(keyValue)
+
